minor cleanup
* move attachment tests into their own file
* fix one code snippet
* stop setting POST manually (its the default)
* dont cast an options object internally when its not necessary
* switch done() with fail() when the block shouldnt be entered
